move project to scalar_nested.cc, add test for different chunking
