+++ cpp/src/arrow/compute/cast.h
@@ -157,17 +157,5 @@ Result<Datum> Cast(const Datum& value, std::shared_ptr<DataType> to_type,
                    const CastOptions& options = CastOptions::Safe(),
                    ExecContext* ctx = NULLPTR);
 
-/// \addtogroup compute-concrete-options
-/// @{
-
-struct ARROW_EXPORT ProjectOptions : public FunctionOptions {
-  explicit ProjectOptions(std::vector<std::string> n) : field_names(std::move(n)) {}
-
-  /// Names for wrapped columns
-  std::vector<std::string> field_names;
-};
-
-/// @}
-
 }  // namespace compute
 }  // namespace arrow
