+++ cpp/src/arrow/compute/kernels/CMakeLists.txt
@@ -25,7 +25,6 @@ add_arrow_compute_test(scalar_test
                        scalar_cast_test.cc
                        scalar_compare_test.cc
                        scalar_nested_test.cc
-                       scalar_project_test.cc
                        scalar_set_lookup_test.cc
                        scalar_string_test.cc
                        scalar_validity_test.cc
